+++ analysis/config_template.ini
@@ -2,10 +2,10 @@
 # Use state abbreviations for the sections.
 [NC]
 DB_NAME = # Name of database
-CMU_SHP = # CMU boundary shapefile name
+CMU_SHP = # CMU centroid point shapefile name
 CMU_SHP_COL_CMU_NAME = # CMU name column name in "CMU_SHP"
 CMU_SHP_COL_RAIN_IN = # CMU rain threshold in inches column name in "CMU_SHP"
-LEASE_SHP = # Leases boundary shapefile name
+LEASE_SHP = # Leases centroid point shapefile name
 LEASE_SHP_COL_LEASE_ID = # Lease ID column name in "LEASE_SHP"
 LEASE_SHP_COL_CMU_NAME = # CMU column name in "LEASE_SHP"
 LEASE_SHP_COL_RAIN_IN = # Rain threshold in inches column name in "LEASE_SHP"
@@ -28,3 +28,9 @@ DB_USER = # DB username
 DB_PASS = # DB Password
 HOST = # Host address
 PORT = # Port number
+
+[Notification]
+EMAIL_SENDER = # Sender email address
+GMAIL_API_CREDENTIAL = # Gmail API credential (https://developers.google.com/gmail/api/quickstart/python)
+EMAIL_RECEIVER = # Receiver email address: separate by comma for multiple receivers (e.g. address1, address2)
+EMAIL_SUBJECT = # Email title
