--- analysis/environment.yml
@@ -1,18 +1,0 @@
-name: shellcast_analysis
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - python=3.7
-  - pip
-  - gdal
-  - rasterio
-  - pygrib
-  - configparser
-  - pandas
-  - geopandas
-  - PyMySQL
-  - SQLAlchemy
-  - rasterstats
-  - shapely
-  - yaml
