--- analysis/requirements.txt
@@ -1,11 +1,0 @@
-configparser >= 5.0.2
-pandas
-geopandas >= 0.9.0
-SQLAlchemy >= 1.4.7
-PyMySQL >= 1.0.2
-pygrib >= 2.1.4
-.\dependencies\GDAL-3.4.2-cp37-cp37m-win_amd64.whl
-.\dependencies\rasterio-1.2.10-cp37-cp37m-win_amd64.whl
-rasterstats >= 0.17.0
-shapely~=2.0.1
-
