+++ analysis/sqlauthproxy_template.bat
@@ -1,0 +1,1 @@
+cloud-sql-proxy -instances=<<Instance Name>>=tcp:3306
