+++ analysis/src/constants.py
@@ -6,18 +6,9 @@ from pathlib import Path
 ROOT_DIR = Path(__file__).resolve().parent.parent
 PREPROC_DATA_DIR = os.path.join(ROOT_DIR, 'data/preprocs')
 PQPF_DATA_DIR = os.path.join(ROOT_DIR, 'data/pqpf')
-LOGS_DIR = os.path.join(ROOT_DIR, 'logs')
 
 # [ File path]
 CONFIG_INI = os.path.join(ROOT_DIR, 'config.ini')
-LOGGING_NC = os.path.join(ROOT_DIR, 'logging_nc.yaml')
-LOGGING_SC = os.path.join(ROOT_DIR, 'logging_sc.yaml')
-
-# [Email]
-EMAIL_SENDER = 'mshukun@ncsu.edu'
-EMAIL_PASSWORD = os.environ.get('SHELLCAST_GMAIL_APP')
-EMAIL_RECEIVER = ['mshukun@ncsu.edu']
-EMAIL_SUBJECT = 'ShellCast Analysis: Error'
 
 # [ FTP ]
 FTP_URL = 'ftp.wpc.ncep.noaa.gov'
